requirements.txt
asgiref==3.5.2
botocore==1.29.13
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
